Remove VariantSelectOverride
Its only purpose was to determine if a particular variant was selected,
but we can determine that using the variants list and the selected
variant.

This also refactors renderProductVariant into a functional component.
Previously it was a HOC that returned a functional component, but
there's no need for that level of complexity when we can just pass all
the necessary arguments as props. This also serves to remove an instance
of using an array index as a React key (although it was combined with
other data so it wasn't that risky).